backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
requests
beautifulsoup4
langchain
langchain-google-genai
pydantic
mysqlclient

